+++ src/main/java/com/gruzini/tennistico/controllers/JoinMatchController.java
@@ -4,7 +4,7 @@ import com.gruzini.tennistico.domain.enums.NotificationType;
 import com.gruzini.tennistico.exceptions.MatchNotFoundException;
 import com.gruzini.tennistico.exceptions.PlayerNotFoundException;
 import com.gruzini.tennistico.exceptions.WrongMatchStatusException;
-import com.gruzini.tennistico.services.JoinMatchService;
+import com.gruzini.tennistico.services.dtos.JoinMatchService;
 import com.gruzini.tennistico.services.NotificationSenderService;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Controller;
