+++ src/main/java/com/gruzini/tennistico/services/InputScoreService.java
@@ -2,6 +2,8 @@ package com.gruzini.tennistico.services;
 
 import com.gruzini.tennistico.domain.Match;
 import com.gruzini.tennistico.domain.enums.MatchStatus;
+import com.gruzini.tennistico.mappers.ScoreMapper;
+import com.gruzini.tennistico.models.dto.ScoreDTO;
 import com.gruzini.tennistico.services.entities.MatchService;
 import com.gruzini.tennistico.services.score.WinDecider;
 import org.springframework.stereotype.Service;
@@ -12,16 +14,25 @@ public class InputScoreService {
     private final MatchService matchService;
     private final RankingService rankingService;
     private final WinDecider winDecider;
+    private final ScoreMapper scoreMapper;
 
-    public InputScoreService(MatchService matchService, RankingService rankingService, WinDecider winDecider) {
+    public InputScoreService(final MatchService matchService,
+                             final RankingService rankingService,
+                             final WinDecider winDecider,
+                             final ScoreMapper scoreMapper) {
         this.matchService = matchService;
         this.rankingService = rankingService;
         this.winDecider = winDecider;
+        this.scoreMapper = scoreMapper;
     }
 
-    public void inputScore(final Long matchId, final String score) {
+    public void inputScore(final Long matchId, final ScoreDTO scoreDTO) {
         final Match match = matchService.getById(matchId);
+        final String score = scoreMapper.mapScoreToString(scoreDTO);
+        updateMatchAndRanking(match, score);
+    }
 
+    private void updateMatchAndRanking(final Match match, final String score){
         matchService.updateMatchScore(match, score);
 
         rankingService.updateWinner(winDecider.decideWinner(match));
